+++ examples/authentication/antd/src/pages/auth/resetPassword/index.tsx
@@ -1,5 +1,0 @@
-import { AuthPage } from "@pankod/refine-antd";
-
-export const ResetPasswordPage: React.FC = () => {
-    return <AuthPage type="resetPassword" />;
-};